+++ inc/ti/fn/fnindexof.h
@@ -40,6 +40,6 @@ done:
 
 static int do__f_indexof(ti_query_t * query, cleri_node_t * nd, ex_t * e)
 {
-    log_debug("function `indexof` is deprecated, use `index_of` instead");
+    log_warning("function `indexof` is deprecated, use `index_of` instead");
     return do__f_index_of(query, nd, e);
 }