+++ inc/ti/fn/fnisbytes.h
@@ -19,6 +19,6 @@ static int do__f_is_bytes(ti_query_t * query, cleri_node_t * nd, ex_t * e)
 
 static int do__f_isbytes(ti_query_t * query, cleri_node_t * nd, ex_t * e)
 {
-    log_debug("function `isbytes` is deprecated, use `is_bytes` instead");
+    log_warning("function `isbytes` is deprecated, use `is_bytes` instead");
     return do__f_is_bytes(query, nd, e);
 }