+++ inc/ti/fn/fnstartswith.h
@@ -30,7 +30,7 @@ failed:
 
 static int do__f_startswith(ti_query_t * query, cleri_node_t * nd, ex_t * e)
 {
-    log_debug("function `startswith` is deprecated, use `starts_with` instead");
+    log_warning("function `startswith` is deprecated, use `starts_with` instead");
     return do__f_starts_with(query, nd, e);
 }